add remoting module.

+++ fastrpc-remoting/fastrpc-remoting-api/src/main/java/com/fast/fastrpc/exchange/Response.java
@@ -18,7 +18,7 @@ public class Response {
     public static final byte CLIENT_ERROR = 9;
     public static final byte SERVER_POOL_EXHAUSTED_ERROR = 10;
 
-    private long id = 0;
+    private int id = 0;
 
     private String version;
 
@@ -35,11 +35,11 @@ public class Response {
     public Response() {
     }
 
-    public Response(long id) {
+    public Response(int id) {
         this.id = id;
     }
 
-    public Response(long id, String version) {
+    public Response(int id, String version) {
         this.id = id;
         this.version = version;
     }
@@ -48,7 +48,7 @@ public class Response {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
